refactor: user, server, roles commands moved to info command as sub commands
